asdasds

+++ aspnetcore/sellerproto/Domain/Model/PurchaseOrder.cs
@@ -8,11 +8,11 @@ namespace XingZen.Domain.Model
 
         public string SalesPerson { get; }
 
-        public decimal Amount { get; }
+        public double Amount { get; }
 
         public string Currency { get; }
 
-        public PurchaseOrder(string purchaseOrderId, string storeId, string salesPerson, decimal amount, string currency)
+        public PurchaseOrder(string purchaseOrderId, string storeId, string salesPerson, double amount, string currency)
         {
             PurchaseOrderId = purchaseOrderId;
             StoreId = storeId;
